+++ springBoot/src/main/java/TaoTaoPerto/springBoot/usuarios/repository/EnderecoRepository.java
@@ -4,10 +4,9 @@ import TaoTaoPerto.springBoot.usuarios.model.EnderecoModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.Arrays;
 import java.util.List;
 
 @Repository
-public interface EnderecoRepository extends JpaRepository<EnderecoModel, Integer> {
-    List<EnderecoModel> findByUsuarioId(Integer usuarioId);
+public interface EnderecoRepository extends JpaRepository<EnderecoModel, Long> {
+    List<EnderecoModel> findByUsuarioId(Long usuarioId);
 }
